# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrainCoach API",
    "version": "v1"
  },
  "paths": {
...
    "/api/relationships/invite": {
      "post": {
        "tags": [
          "Relationships"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAthleteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAthleteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InviteAthleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationships/{id}/respond": {
      "post": {
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInviteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInviteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationships/{id}/revoke": {
      "post": {
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAccessRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAccessRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationships/{id}/permissions": {
      "put": {
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPermissionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPermissionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetPermissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachAthleteRelationshipDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/reports/{date}/{type}": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReportType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/reports/{date}/{type}/regenerate": {
      "post": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReportType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedReportDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/athletes/{athleteUserId}/sleep": {
      "get": {
        "tags": [
          "SleepRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SleepRecordDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SleepRecordDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SleepRecordDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SleepRecords"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSleepRecordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSleepRecordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSleepRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SleepRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepRecordDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoachAthleteRelationshipDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "coachUserId": {
            "type": "string",
            "format": "uuid"
          },
          "coachName": {
            "type": "string",
            "nullable": true
          },
          "coachEmail": {
            "type": "string",
            "nullable": true
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "athleteName": {
            "type": "string",
            "nullable": true
          },
          "athleteEmail": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/RelationshipStatus"
          },
          "invitedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "respondedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "inviteNote": {
            "type": "string",
            "nullable": true
          },
          "grantedScopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionScope"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataSource": {
        "enum": [
          "Manual",
          "FileImport",
          "Strava",
          "GarminDemoProvider",
          "MySasyDemoProvider"
        ],
        "type": "string"
      },
...
      "GeneratedReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "narrativeText": {
            "type": "string",
            "nullable": true
          },
          "deliveryChannel": {
            "$ref": "#/components/schemas/ReportDeliveryChannel"
          },
          "deliveryStatus": {
            "$ref": "#/components/schemas/ReportDeliveryStatus"
          },
          "insights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeneratedReportInsightDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeneratedReportInsightDto": {
        "type": "object",
        "properties": {
          "ruleCode": {
            "type": "string",
            "nullable": true
          },
          "severity": {
            "$ref": "#/components/schemas/InsightSeverity"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InsightSeverity": {
        "enum": [
          "Info",
          "Notice",
          "Attention"
        ],
        "type": "string"
      },
...
      "InviteAthleteRequest": {
        "type": "object",
        "properties": {
          "athleteEmail": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PermissionScope": {
        "enum": [
          "ViewTrainingPlan",
          "EditTrainingPlan",
          "ViewCompletedActivities",
          "ViewWellness",
          "ViewNutrition",
          "ViewHealthFlags",
          "CommentOnWorkouts",
          "ViewPersonalRecords"
        ],
        "type": "string"
      },
...
      "RelationshipStatus": {
        "enum": [
          "PendingInvite",
          "Active",
          "Revoked",
          "Ended",
          "Declined"
        ],
        "type": "string"
      },
      "ReportDeliveryChannel": {
        "enum": [
          "InApp",
          "Email",
          "Push",
          "Telegram"
        ],
        "type": "string"
      },
      "ReportDeliveryStatus": {
        "enum": [
          "Pending",
          "Delivered",
          "Failed",
          "NotConfigured"
        ],
        "type": "string"
      },
      "ReportType": {
        "enum": [
          "Morning",
          "Evening"
        ],
        "type": "string"
      },
      "RespondToInviteRequest": {
        "type": "object",
        "properties": {
          "accept": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RevokeAccessRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SetPermissionRequest": {
        "type": "object",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/PermissionScope"
          },
          "granted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SleepRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deepSleepMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "remSleepMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpsertSleepRecordRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deepSleepMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "remSleepMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Zadejte JWT access token (bez prefixu \"Bearer \").",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}